locale.c: Skip changing the locale if same as the old
Meet the new boss; same as the old boss.

There is some housekeeping involved in changing locales, particularly in
computing a new collation ordering.  So, it's worth it to check if the
new locale being requested is actually not different than the existing.
If so, we can just return without doing anything